+++ src/Symfony/Component/Messenger/Handler/HandlerDescriptor.php
@@ -19,12 +19,34 @@ namespace Symfony\Component\Messenger\Handler;
 final class HandlerDescriptor
 {
     private $handler;
+    private $name;
+    private $batchHandler;
     private $options;
 
     public function __construct(callable $handler, array $options = [])
     {
+        if (!$handler instanceof \Closure) {
+            $handler = \Closure::fromCallable($handler);
+        }
+
         $this->handler = $handler;
         $this->options = $options;
+
+        $r = new \ReflectionFunction($handler);
+
+        if (str_contains($r->name, '{closure}')) {
+            $this->name = 'Closure';
+        } elseif (!$handler = $r->getClosureThis()) {
+            $class = $r->getClosureScopeClass();
+
+            $this->name = ($class ? $class->name.'::' : '').$r->name;
+        } else {
+            if ($handler instanceof BatchHandlerInterface) {
+                $this->batchHandler = $handler;
+            }
+
+            $this->name = \get_class($handler).'::'.$r->name;
+        }
     }
 
     public function getHandler(): callable
@@ -34,7 +56,7 @@ final class HandlerDescriptor
 
     public function getName(): string
     {
-        $name = $this->callableName($this->handler);
+        $name = $this->name;
         $alias = $this->options['alias'] ?? null;
 
         if (null !== $alias) {
@@ -44,37 +66,13 @@ final class HandlerDescriptor
         return $name;
     }
 
+    public function getBatchHandler(): ?BatchHandlerInterface
+    {
+        return $this->batchHandler;
+    }
+
     public function getOption(string $option)
     {
         return $this->options[$option] ?? null;
-    }
-
-    private function callableName(callable $handler): string
-    {
-        if (\is_array($handler)) {
-            if (\is_object($handler[0])) {
-                return \get_class($handler[0]).'::'.$handler[1];
-            }
-
-            return $handler[0].'::'.$handler[1];
-        }
-
-        if (\is_string($handler)) {
-            return $handler;
-        }
-
-        if ($handler instanceof \Closure) {
-            $r = new \ReflectionFunction($handler);
-            if (str_contains($r->name, '{closure}')) {
-                return 'Closure';
-            }
-            if ($class = $r->getClosureScopeClass()) {
-                return $class->name.'::'.$r->name;
-            }
-
-            return $r->name;
-        }
-
-        return \get_class($handler).'::__invoke';
     }
 }