+++ src/Symfony/Component/PropertyInfo/CHANGELOG.md
@@ -7,6 +7,11 @@ CHANGELOG
  * Remove the `Type::getCollectionKeyType()` and `Type::getCollectionValueType()` methods, use `Type::getCollectionKeyTypes()` and `Type::getCollectionValueTypes()` instead
  * Remove the `enable_magic_call_extraction` context option in `ReflectionExtractor::getWriteInfo()` and `ReflectionExtractor::getReadInfo()` in favor of `enable_magic_methods_extraction`
 
+5.4
+---
+
+ * Add PhpStanExtractor
+
 5.3
 ---